chore(domain): ETL box file separation management
